Renamed files and functions, invalid heredoc for cases with outfile, heredoc and command
