+++ src/builtins/pwd.c
@@ -1,12 +1,12 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   lexer.c                                            :+:      :+:    :+:   */
+/*   pwd.c                                              :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: plouda <marvin@42.fr>                      +#+  +:+       +#+        */
+/*   By: plouda <plouda@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/06 14:53:16 by plouda            #+#    #+#             */
-/*   Updated: 2023/07/19 14:08:04 by plouda           ###   ########.fr       */
+/*   Updated: 2023/08/25 10:25:41 by plouda           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,8 @@ int	pwd_built(char **env)
 {
 	char	*cwd;
 
-	env = env;
+	//env = env;
+	(void)env;
 	cwd = malloc(1000);
 	getcwd(cwd, 1000);
 	ft_putstr_fd(cwd, 1);
